+++ .onchange.sh
@@ -8,6 +8,7 @@ nameonly="${filename%.*}"
 case $extension in
     go)
         goimports -w $path
+	go generate .
         ;;
 esac
 go test -coverprofile /tmp/c.out ./...