+++ editorut_test.go
@@ -141,35 +141,3 @@ func (me *EditorUnderTest) mustExport(w io.Writer) {
 		me.T.Fatal(err)
 	}
 }
-
-func (me *EditorUnderTest) shouldSaveReport(filename string, report Report) {
-	err := me.SaveReport(filename, report)
-	if err != nil {
-		me.T.Helper()
-		me.T.Error(err)
-	}
-}
-
-func (me *EditorUnderTest) mustSaveReport(filename string, report Report) {
-	err := me.SaveReport(filename, report)
-	if err != nil {
-		me.T.Helper()
-		me.T.Fatal(err)
-	}
-}
-
-func (me *EditorUnderTest) shouldWriteReport(w io.Writer, report Report) {
-	err := me.WriteReport(w, report)
-	if err != nil {
-		me.T.Helper()
-		me.T.Error(err)
-	}
-}
-
-func (me *EditorUnderTest) mustWriteReport(w io.Writer, report Report) {
-	err := me.WriteReport(w, report)
-	if err != nil {
-		me.T.Helper()
-		me.T.Fatal(err)
-	}
-}